Reorganize source files by feature
