+++ src/index.html
@@ -19,6 +19,6 @@
             <span id="tip">Tap slot to build. Tap tower to switch (1 gold).</span>
         </div>
         <canvas id="game" width="450" height="800"></canvas>
-        <script type="module" src="main.js"></script>
+        <script type="module" src="js/main.js"></script>
     </body>
 </html>
