+++ src/js/systems/assets.js
